Renamed siaMath to hostMath and reorganized code
