+++ plugins/Hosting/js/host.js
@@ -3,9 +3,9 @@
 // Manges the lifecycle of the plugin
 function host() {
 	// Tracks details about the various hosting properties
-	var props = require('./hostData.js');
+	var props = require('./hostProperties.js');
 	// Hold Sia math logic 
-	var S = require('./siaMath.js');
+	var math = require('./hostMath.js');
 	// Names of configurable host properties
 	var configurable = ['totalstorage', 'price'];
 	// Tracks if props have been made
@@ -29,7 +29,7 @@ function host() {
 			`);
 			el.attr('id', name);
 			el.find('.name').text(prop.descr + ' (' + prop.unit + ')');
-			el.find('.value').text(S.convert(prop));
+			el.find('.value').text(math.convertProperty(prop));
 			$('#properties').append(el);
 		});
 	}
@@ -42,13 +42,13 @@ function host() {
 		});
 
 		// Process host info
-		var totalStorage = S.formatByte(props.totalstorage);
-		var usedStorage = S.formatSiacoin(props.totalstorage.value - props.storageremaining.value);
-		var revenue = S.formatSiacoin(props.revenue);
-		var upcomingRevenue = S.formatSiacoin(props.upcomingrevenue);
+		var totalStorage = math.formatByte(props.totalstorage.value);
+		var usedStorage = math.formatByte(props.totalstorage.value - props.storageremaining.value);
+		var revenue = math.formatSiacoin(props.revenue.value);
+		var upcomingRevenue = math.formatSiacoin(props.upcomingrevenue.value);
 
 		// Update host info
-		$('#numcontracts').text(S.format(props.numcontracts));
+		$('#numcontracts').text(math.formatProperty(props.numcontracts));
 		$('#storage').text(usedStorage + '/' + totalStorage + ' in use');
 		$('#revenue').text(revenue + ' earned');
 		$('#upcomingrevenue').text(upcomingRevenue + ' to be earned');
@@ -76,7 +76,7 @@ function host() {
 	function save(settings) {
 		// Revert each value to base units
 		$.each(settings, function(name, value) {
-			settings[name] = S.revert(value, props[name].conversion);
+			settings[name] = math.revertToBaseUnit(value, props[name].conversion);
 		});
 		// Send configuration call
 		Daemon.apiCall({
@@ -93,7 +93,7 @@ function host() {
 	function resetValues() {
 		var convertedValues = {};
 		configurable.forEach(function(name) {
-			convertedValues[name] = S.convert(props[name]);
+			convertedValues[name] = math.convertProperty(props[name]);
 		});
 		return convertedValues;
 	}
